Back | new userOnProject relation, project, board entities changes

+++ track_it_back_end/src/board/board.service.ts
@@ -2,16 +2,18 @@ import { BadRequestException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { BoardDto } from './dto/board.dto';
 import { ProjectService } from 'src/project/project.service';
+import { UserOnProjectService } from 'src/user-on-project/user-on-project.service';
 
 @Injectable()
 export class BoardService {
 	constructor(
 		private prisma: PrismaService,
-		private projectService: ProjectService
+		private projectService: ProjectService,
+		private userOnProject: UserOnProjectService
 	) {}
 
-	async getAll(createdBy: string) {
-		return this.projectService.getBoards(createdBy);
+	async getAll(projectId: string) {
+		return this.projectService.getBoards(projectId);
 	}
 
 	async getColumns(boardId: string) {
@@ -26,7 +28,13 @@ export class BoardService {
 		});
 	}
 
-	async create(dto: BoardDto, projectId: string) {
+	async create(dto: BoardDto, projectId: string, userId: string) {
+		const user = this.userOnProject.getById(userId, projectId);
+
+		if (!user)
+			throw new BadRequestException(
+				'user must be connected to this project to do this'
+			);
 		return this.prisma.board.create({
 			data: {
 				...dto,
@@ -39,7 +47,17 @@ export class BoardService {
 		});
 	}
 
-	async update(dto: BoardDto, boardId: string) {
+	async update(
+		dto: BoardDto,
+		boardId: string,
+		userId: string,
+		projectId: string
+	) {
+		const user = await this.userOnProject.getById(userId, projectId);
+		if (!user)
+			throw new BadRequestException(
+				'user must be connected to this project to do this'
+			);
 		return this.prisma.board.update({
 			where: {
 				id: boardId,
@@ -48,7 +66,12 @@ export class BoardService {
 		});
 	}
 
-	async delete(boardId: string) {
+	async delete(boardId: string, userId: string, projectId: string) {
+		const admin = await this.userOnProject.getAdmin(userId, projectId);
+
+		if (!admin)
+			throw new BadRequestException('you must be an admin to do this');
+
 		return this.prisma.board.delete({
 			where: {
 				id: boardId,
