+++ track_it_back_end/src/project/project.service.ts
@@ -1,17 +1,33 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { ProjectDto } from './dto/project.dto';
+import { UserOnProjectService } from 'src/user-on-project/user-on-project.service';
 
 @Injectable()
 export class ProjectService {
-	constructor(private prisma: PrismaService) {}
+	constructor(
+		private prisma: PrismaService,
+		private userOnProject: UserOnProjectService
+	) {}
 
 	async getAll(createdBy: string) {
 		return this.prisma.project.findMany({
 			where: {
 				createdBy,
 			},
+			include: {
+				users: {
+					select: {
+						userId: true,
+						role: true,
+					},
+				},
+			},
 		});
+	}
+
+	async getAllUserOnProject() {
+		return this.userOnProject.getAll();
 	}
 
 	async getAllDetailed(createdBy: string) {
@@ -38,10 +54,10 @@ export class ProjectService {
 		});
 	}
 
-	async getBoards(createdBy: string) {
+	async getBoards(projectId: string) {
 		return this.prisma.project.findMany({
 			where: {
-				createdBy,
+				id: projectId,
 			},
 			select: {
 				id: true,
@@ -51,30 +67,68 @@ export class ProjectService {
 		});
 	}
 
-	async create(dto: ProjectDto, createdBy: string) {
-		return this.prisma.project.create({
+	async create(dto: ProjectDto, userId: string) {
+		const project = await this.prisma.project.create({
 			data: {
 				...dto,
 				user: {
 					connect: {
-						id: createdBy,
+						id: userId,
 					},
 				},
 			},
 		});
+		const projectId = project.id;
+		this.userOnProject.createAdmin(userId, projectId);
+		return project;
 	}
 
-	async update(dto: ProjectDto, projectId: string, createdBy: string) {
+	async connect(userId: string, projectId: string) {
+		const oldUser = await this.userOnProject.getById(userId, projectId);
+
+		if (oldUser)
+			throw new BadRequestException('User already connected to this project');
+
+		return this.prisma.userOnProject.create({
+			data: {
+				userId: userId,
+				projectId: projectId,
+				role: 'editor',
+			},
+		});
+	}
+
+	async disconnect(userId: string, projectId: string) {
+		const oldUser = await this.userOnProject.getById(userId, projectId);
+
+		if (!oldUser)
+			throw new BadRequestException('User do not exist in this project');
+		if (oldUser.role == 'admin')
+			throw new BadRequestException("You can't leave this project as creator");
+
+		return this.userOnProject.delete(userId, projectId);
+	}
+
+	async update(dto: ProjectDto, projectId: string, userId: string) {
+		const admin = await this.userOnProject.getAdmin(userId, projectId);
+
+		if (!admin)
+			throw new BadRequestException('you must be an admin to do this');
+
 		return this.prisma.project.update({
 			where: {
-				createdBy,
 				id: projectId,
 			},
 			data: dto,
 		});
 	}
 
-	async delete(projectId: string) {
+	async delete(projectId: string, userId: string) {
+		const admin = await this.userOnProject.getAdmin(userId, projectId);
+
+		if (!admin)
+			throw new BadRequestException('you must be an admin to do this');
+
 		return this.prisma.project.delete({
 			where: {
 				id: projectId,
